+++ packages/connection-storage/src/connection-repository.ts
@@ -1,7 +1,5 @@
-import { createElement, useMemo, useContext } from 'react';
-import { createContext } from 'react';
 import type { ConnectionInfo } from '@mongodb-js/connection-info';
-import { type ConnectionStorage, connectionStorageLocator } from './renderer';
+import { type ConnectionStorage } from './renderer';
 import ConnectionString from 'mongodb-connection-string-url';
 import { merge } from 'lodash';
 
@@ -62,30 +60,4 @@ export class ConnectionRepository {
     const bName = b.favorite?.name?.toLocaleLowerCase() || '';
     return aName.localeCompare(bName);
   };
-}
-
-export const ConnectionRepositoryContext =
-  createContext<ConnectionRepository | null>(null);
-
-export const ConnectionRepositoryContextProvider: React.FunctionComponent<
-  object
-> = ({ children }) => {
-  const storage = connectionStorageLocator();
-  const value = useMemo(() => new ConnectionRepository(storage), [storage]);
-
-  return createElement(ConnectionRepositoryContext.Provider, {
-    value,
-    children,
-  });
-};
-
-export function connectionRepositoryLocator(): ConnectionRepository {
-  const connectionRepository = useContext(ConnectionRepositoryContext);
-  if (!connectionRepository) {
-    throw new Error(
-      'Could not find the current ConnectionRepository. Did you forget to setup the ConnectionRepositoryContext?'
-    );
-  }
-
-  return connectionRepository;
 }
