+++ packages/connection-storage/src/main.ts
@@ -1,12 +1,6 @@
 export type { ConnectionInfo } from '@mongodb-js/connection-info';
-export {
-  ConnectionRepository,
-  ConnectionRepositoryContext,
-  ConnectionRepositoryContextProvider,
-  connectionRepositoryLocator,
-} from './connection-repository';
+export { ConnectionRepository } from './connection-repository';
 export { ConnectionStorage } from './connection-storage';
-export { ConnectionStorageContext, connectionStorageLocator } from './renderer';
 export type {
   ExportConnectionOptions,
   ImportConnectionOptions,
